fix(playwright-vscode-ext): eliminate circular dependencies - W-22862555
