+++ packages/playwright-vscode-ext/test/playwright/specs/fileOperations.headless.spec.ts
@@ -7,12 +7,8 @@
 
 import { expect } from '@playwright/test';
 import { createFileWithContents } from '../../../src/utils/fileHelpers';
-import {
-  waitForVSCodeWorkbench,
-  closeWelcomeTabs,
-  waitForWorkspaceReady,
-  ensureSecondarySideBarHidden
-} from '../../../src/utils/helpers';
+import { waitForVSCodeWorkbench, closeWelcomeTabs, waitForWorkspaceReady } from '../../../src/utils/helpers';
+import { ensureSecondarySideBarHidden } from '../../../src/utils/workflows';
 import { EDITOR_WITH_URI, TAB } from '../../../src/utils/locators';
 import { test } from '../fixtures/index';
